add smooth scroll
